# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Elevate API",
    "version": "1.0.0",
    "description": "REST API for Elevate, a dark-themed Kanban and planning app. Authenticate via JWT bearer tokens (from /api/v1/auth/login) or API keys (X-API-Key header or ?api_key= query param). Use workspace_id to scope all operations.",
    "contact": {
      "name": "Elevate"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "API v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "key_prefix": {
            "type": "string"
          },
          "scopes": {
            "type": "string"
          },
          "last_used_at": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "example": "2026-05-16T10:30:00.000Z"
          },
          "expires_at": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "example": "2026-05-16T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "key_prefix",
          "scopes"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/workspaces/{workspaceId}/api-keys": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "List API keys",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "workspaceId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "API key list"
          },
          "403": {
            "description": "User session required"
          }
        }
      },
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Create API key",
        "description": "The raw key value is returned only once at creation time. Viewers cannot create API keys.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "workspaceId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "scopes": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "expires_at": {
                    "type": "string",
                    "maxLength": 64
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiKey"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string",
                          "description": "Returned only once. Store securely."
                        },
                        "created_at": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "key",
                        "created_at"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "User session required"
          }
        }
      }
    },
    "/workspaces/{workspaceId}/api-keys/{keyId}": {
      "delete": {
        "tags": [
          "API Keys"
        ],
        "summary": "Revoke API key",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "workspaceId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "keyId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Revoked"
          },
          "403": {
            "description": "User session required"
          },
          "404": {
            "description": "API key not found"
          }
        }
      }
    },
    "/tasks/{taskId}/attachments": {
      "post": {
        "tags": [
          "Attachments"
        ],
        "summary": "Upload attachment",
        "description": "Image upload (JPEG/PNG/GIF/WebP), max 5 MB. The response carries the public URL the client should use.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "taskId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "task_id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "mime_type": {
                      "type": "string",
                      "nullable": true
                    },
                    "size": {
                      "type": "number",
                      "nullable": true
                    },
                    "sha256": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "task_id",
                    "type",
                    "url",
                    "name"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid file"
          }
        }
      }
    },
    "/attachments/{id}/file": {
      "get": {
        "tags": [
          "Attachments"
        ],
        "summary": "Download attachment file",
        "description": "Streams the attachment bytes. Authenticated via either the standard session/API-key headers, or via a signed `?token=` query parameter (used by `<img>` tags which cannot send headers). Tokens expire after 1 hour and are bound to the requesting user.",
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "token",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "File stream"
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "404": {
            "description": "Attachment not found"
          }
        }
      }
    },
    "/attachments/{id}": {
      "delete": {
        "tags": [
          "Attachments"
        ],
        "summary": "Delete attachment",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
...
  }
}